refactor(moveRule): 클래스명 변경, 메서드명 변경
