+++ src/main/java/janggi/domain/moveRule/MoveRule.java
@@ -3,18 +3,18 @@ package janggi.domain.moveRule;
 import janggi.domain.board.PiecePath;
 import janggi.domain.board.Position;
 import janggi.domain.moveRule.pathStrategy.PathStrategy;
-import janggi.domain.moveRule.moveStrategy.moveStrategy;
+import janggi.domain.moveRule.routeStrategy.RouteStrategy;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.TeamColor;
 import java.util.List;
 
 public class MoveRule {
     private final PathStrategy pathStrategy;
-    private final moveStrategy moveStrategy;
+    private final RouteStrategy RouteStrategy;
 
-    public MoveRule(PathStrategy pathStrategy, moveStrategy moveStrategy) {
+    public MoveRule(PathStrategy pathStrategy, RouteStrategy RouteStrategy) {
         this.pathStrategy = pathStrategy;
-        this.moveStrategy = moveStrategy;
+        this.RouteStrategy = RouteStrategy;
     }
 
     public boolean verifyMovement(PiecePath path, TeamColor teamColor) {
@@ -22,10 +22,10 @@ public class MoveRule {
     }
 
     public List<Position> findAllRoute(PiecePath path) {
-        return pathStrategy.findAllRoute(path);
+        return pathStrategy.findAllIntermediatePositions(path);
     }
 
     public boolean verifyRoute(Piece piece, Piece destinationPiece, List<Piece> allPiecesOnRoute) {
-        return moveStrategy.canMoveAlongRoute(piece, destinationPiece, allPiecesOnRoute);
+        return RouteStrategy.canMoveAlongRoute(piece, destinationPiece, allPiecesOnRoute);
     }
 }
