+++ src/main/java/janggi/domain/moveRule/pathStrategy/PalacePathStrategy.java
@@ -30,7 +30,7 @@ public class PalacePathStrategy implements PathStrategy {
     }
 
     @Override
-    public List<Position> findAllRoute(PiecePath path) {
+    public List<Position> findAllIntermediatePositions(PiecePath path) {
         return List.of();
     }
 }
