+++ src/main/java/janggi/domain/moveRule/pathStrategy/SoldierPathStrategy.java
@@ -52,7 +52,7 @@ public class SoldierPathStrategy implements PathStrategy {
     }
 
     @Override
-    public List<Position> findAllRoute(PiecePath path) {
+    public List<Position> findAllIntermediatePositions(PiecePath path) {
         return List.of();
     }
 }
