+++ src/main/java/janggi/domain/moveRule/routeStrategy/RouteStrategy.java
@@ -1,8 +1,8 @@
-package janggi.domain.moveRule.moveStrategy;
+package janggi.domain.moveRule.routeStrategy;
 
 import janggi.domain.piece.Piece;
 import java.util.List;
 
-public interface moveStrategy {
+public interface RouteStrategy {
     boolean canMoveAlongRoute(Piece piece, Piece destination, List<Piece> piecesInRoute);
 }
