+++ src/main/java/janggi/domain/piece/PieceType.java
@@ -6,17 +6,17 @@ import janggi.domain.moveRule.pathStrategy.HorsePathStrategy;
 import janggi.domain.moveRule.pathStrategy.PalacePathStrategy;
 import janggi.domain.moveRule.pathStrategy.SoldierPathStrategy;
 import janggi.domain.moveRule.pathStrategy.StraightPathStrategy;
-import janggi.domain.moveRule.moveStrategy.CannonMoveStrategy;
-import janggi.domain.moveRule.moveStrategy.DefaultMoveStrategy;
+import janggi.domain.moveRule.routeStrategy.CannonRouteStrategy;
+import janggi.domain.moveRule.routeStrategy.DefaultRouteStrategy;
 
 public enum PieceType {
-    SOLDIER(2, new MoveRule(new SoldierPathStrategy(), new DefaultMoveStrategy())),
-    HORSE(5, new MoveRule(new HorsePathStrategy(), new DefaultMoveStrategy())),
-    CHARIOT(13, new MoveRule(new StraightPathStrategy(), new DefaultMoveStrategy())),
-    ELEPHANT(3, new MoveRule(new ElephantPathStrategy(), new DefaultMoveStrategy())),
-    CANNON(7, new MoveRule(new StraightPathStrategy(),new  CannonMoveStrategy())),
-    GENERAL(0, new MoveRule(new PalacePathStrategy(), new DefaultMoveStrategy())),
-    GUARD(3, new MoveRule(new PalacePathStrategy(), new DefaultMoveStrategy())),
+    SOLDIER(2, new MoveRule(new SoldierPathStrategy(), new DefaultRouteStrategy())),
+    HORSE(5, new MoveRule(new HorsePathStrategy(), new DefaultRouteStrategy())),
+    CHARIOT(13, new MoveRule(new StraightPathStrategy(), new DefaultRouteStrategy())),
+    ELEPHANT(3, new MoveRule(new ElephantPathStrategy(), new DefaultRouteStrategy())),
+    CANNON(7, new MoveRule(new StraightPathStrategy(),new CannonRouteStrategy())),
+    GENERAL(0, new MoveRule(new PalacePathStrategy(), new DefaultRouteStrategy())),
+    GUARD(3, new MoveRule(new PalacePathStrategy(), new DefaultRouteStrategy())),
     NONE(0, null)
     ;
